refactor(tests): merge deps tests into existing files
Per CONTRIBUTING.md: "Do NOT create new test files." Moved 20 tests
from tests/core/deps.test.ts → tests/store.test.ts (ContentStore, manifest
mutations, dep opening) and 2 tests from tests/core/deps-integration.test.ts
→ tests/core/search.test.ts (federated search pipeline). No new files.
