+++ tests/core/search.test.ts
@@ -14,13 +14,14 @@
 import { describe, test, expect, it, beforeEach, afterEach } from "vitest";
 import { strict as assert } from "node:assert";
 import { tmpdir } from "node:os";
-import { join } from "node:path";
+import { join, resolve } from "node:path";
 import { mkdirSync, writeFileSync, rmSync } from "node:fs";
 import { randomUUID } from "node:crypto";
 import { ContentStore } from "../../src/store.js";
 import { SessionDB } from "../../src/session/db.js";
 import { searchAllSources, type UnifiedSearchResult } from "../../src/search/unified.js";
 import { searchAutoMemory } from "../../src/search/auto-memory.js";
+import { resolveDepManifest, computeDepDBPath, openDepStore } from "../../src/deps.js";
 import { extractSnippet, formatBatchQueryResults, positionsFromHighlight } from "../../src/server.js";
 
 // ─────────────────────────────────────────────────────────
@@ -3412,3 +3413,90 @@ describe("searchAllSources with dep stores", () => {
 // ═══════════════════════════════════════════════════════════
 // 11. Knowledge-reuse event (removed — read path must not mutate state)
 // ═══════════════════════════════════════════════════════════
+
+// ═══════════════════════════════════════════════════════════
+// 12. Cross-project dependency search — merged from tests/core/deps-integration.test.ts
+// ═══════════════════════════════════════════════════════════
+
+describe("ctx-deps integration", () => {
+  const TEST_ROOT = join(tmpdir(), "ctx-deps-integration-" + Date.now());
+
+  function cleanup() {
+    try { rmSync(TEST_ROOT, { recursive: true, force: true }); } catch {}
+  }
+
+  beforeEach(cleanup);
+  afterEach(cleanup);
+
+  it("parses manifest, opens dep store, searches across both", () => {
+    const currentDir = join(TEST_ROOT, "current");
+    const depDir = join(TEST_ROOT, "dep");
+    mkdirSync(currentDir, { recursive: true });
+    mkdirSync(depDir, { recursive: true });
+
+    writeFileSync(
+      join(currentDir, ".ctx-deps.json"),
+      JSON.stringify({ dependencies: { "my-dep": { path: "../dep" } } }, null, 2),
+    );
+
+    const configPath = join(tmpdir(), ".claude-test-int");
+    const depDBPath = computeDepDBPath(depDir, configPath);
+    mkdirSync(join(depDBPath, ".."), { recursive: true });
+    const depStore = new ContentStore(depDBPath);
+    depStore.index({
+      content: "# My Dep API\n\n## Functions\n\n- `foo(bar: string): void`\n- `baz(): number`",
+      source: "dep-docs",
+    });
+    depStore.close();
+
+    const currentStore = new ContentStore();
+    currentStore.index({
+      content: "# Current Project\n\nCalls foo() from my-dep",
+      source: "current",
+    });
+
+    const manifest = resolveDepManifest(currentDir);
+    expect(manifest).not.toBeNull();
+
+    const depPath = resolve(currentDir, manifest!.dependencies["my-dep"].path);
+    const openedDepStore = openDepStore(depPath, configPath);
+    expect(openedDepStore).not.toBeNull();
+
+    const results = searchAllSources({
+      query: "foo bar baz", limit: 10, store: currentStore,
+      depStores: new Map([["my-dep", openedDepStore!]]),
+    });
+
+    const depResults = results.filter(r => r.origin === "upstream-dep");
+    expect(results.filter(r => r.origin === "current-session").length).toBeGreaterThan(0);
+    expect(depResults.length).toBeGreaterThan(0);
+    expect(depResults[0].title).toContain("My Dep API");
+  });
+
+  it("gracefully handles missing dep ContentStore (null from openDepStore)", () => {
+    const currentDir = join(TEST_ROOT, "no-dep-db");
+    const depDir = join(TEST_ROOT, "no-dep-db-dep");
+    mkdirSync(currentDir, { recursive: true });
+    mkdirSync(depDir, { recursive: true });
+
+    writeFileSync(
+      join(currentDir, ".ctx-deps.json"),
+      JSON.stringify({ dependencies: { "missing-dep": { path: "../no-dep-db-dep" } } }, null, 2),
+    );
+
+    const configPath = join(tmpdir(), ".claude-test-missing");
+    const manifest = resolveDepManifest(currentDir);
+    expect(manifest).not.toBeNull();
+    expect(openDepStore(depDir, configPath)).toBeNull();
+
+    const currentStore = new ContentStore();
+    currentStore.index({ content: "# Just current", source: "current" });
+
+    const results = searchAllSources({
+      query: "current", limit: 10, store: currentStore,
+      depStores: new Map(),
+    });
+    expect(results.length).toBeGreaterThan(0);
+    expect(results.every(r => r.origin === "current-session")).toBe(true);
+  });
+});
